fix: move tests to correct location, fix React imports
- Move badge tests from src/components/ui/ to test/components/
- Fix React import to use namespace import pattern
- Remove JSX.Element return type annotation
