+++ packages/ui/src/components/ui/badge.tsx
@@ -1,7 +1,7 @@
-import type { HTMLAttributes } from 'react';
+import type * as React from 'react';
 import classy from '../../primitives/classy';
 
-export interface BadgeProps extends HTMLAttributes<HTMLSpanElement> {
+export interface BadgeProps extends React.HTMLAttributes<HTMLSpanElement> {
   /** Visual variant */
   variant?: 'default' | 'secondary' | 'destructive' | 'success' | 'warning' | 'info' | 'outline';
 }
@@ -16,14 +16,11 @@ const variantClasses: Record<string, string> = {
   outline: 'bg-transparent border border-input text-foreground',
 };
 
-export function Badge({ className, variant = 'default', ...props }: BadgeProps): JSX.Element {
-  const baseClasses = 'inline-flex items-center justify-center rounded-full px-2.5 py-0.5 text-xs font-semibold transition-colors';
-  
-  const classes = classy(
-    baseClasses,
-    variantClasses[variant] ?? variantClasses.default,
-    className,
-  );
+export function Badge({ className, variant = 'default', ...props }: BadgeProps) {
+  const baseClasses =
+    'inline-flex items-center justify-center rounded-full px-2.5 py-0.5 text-xs font-semibold transition-colors';
+
+  const classes = classy(baseClasses, variantClasses[variant] ?? variantClasses.default, className);
 
   return <span className={classes} {...props} />;
 }
